@@ -1,4 +1,4 @@
-class AddProviderAndUidToUsers < ActiveRecord::Migration
+class AddProviderAndUidToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :provider, :string
     add_index :users, :provider
