@@ -1,4 +1,4 @@
-class AddTrackingModeToEstimations < ActiveRecord::Migration
+class AddTrackingModeToEstimations < ActiveRecord::Migration[4.2]
   def change
     add_column :estimations, :tracking_mode, :boolean, null: false, default: false
   end
