@@ -2,7 +2,7 @@ require 'rails_helper'
 
 describe EstimationDecorator do
   it "rounds large numbers as integers" do
-    estimation = FactoryGirl.create(:estimation_with_items, items: {count: 4, size: 100}).decorate
+    estimation = FactoryBot.create(:estimation_with_items, items: {count: 4, size: 100}).decorate
 
     expect(estimation.sum).to eq("400")
     expect(estimation.buffer).to eq("200")
@@ -10,7 +10,7 @@ describe EstimationDecorator do
   end
 
   it "rounds one-digit values to .XX" do
-    estimation = FactoryGirl.create(:estimation_with_items, items: {count: 2, size: 1}).decorate
+    estimation = FactoryBot.create(:estimation_with_items, items: {count: 2, size: 1}).decorate
 
     expect(estimation.sum).to eq("2")
     expect(estimation.buffer).to eq("1.41")
@@ -18,7 +18,7 @@ describe EstimationDecorator do
   end
 
   it 'works with empty estimations' do
-    estimation = FactoryGirl.create(:estimation).decorate
+    estimation = FactoryBot.create(:estimation).decorate
 
     expect(estimation.sum).to eq("0")
     expect(estimation.buffer).to eq("0")
@@ -26,7 +26,7 @@ describe EstimationDecorator do
   end
 
   it "casts round floats to integers" do
-    estimation = FactoryGirl.create(:estimation_with_items, items: {count: 1, size: 1}).decorate
+    estimation = FactoryBot.create(:estimation_with_items, items: {count: 1, size: 1}).decorate
 
     expect(estimation.sum).to eq("1")
     expect(estimation.buffer).to eq("1")
@@ -35,21 +35,21 @@ describe EstimationDecorator do
 
   describe '#items_partial_name' do
     it 'returns correct partial for estimation mode' do
-      estimation = FactoryGirl.create(:estimation).decorate
+      estimation = FactoryBot.create(:estimation).decorate
       expect(estimation.items_partial_name).to eq("estimation_items/estimation_item")
     end
 
     it 'returns correct for tracking mode' do
-      estimation = FactoryGirl.create(:estimation, tracking_mode: true).decorate
+      estimation = FactoryBot.create(:estimation, tracking_mode: true).decorate
       expect(estimation.items_partial_name).to eq("estimation_items/estimation_item_trackable")
     end
   end
 
   describe '#actual_sum' do
     it 'returns correct partial for estimation mode' do
-      estimation = FactoryGirl.create(:estimation)
-      FactoryGirl.create :estimation_item, estimation: estimation, actual_value: 1
-      FactoryGirl.create :estimation_item, estimation: estimation
+      estimation = FactoryBot.create(:estimation)
+      FactoryBot.create :estimation_item, estimation: estimation, actual_value: 1
+      FactoryBot.create :estimation_item, estimation: estimation
 
       expect(estimation.decorate.actual_sum).to eq(1.0)
     end
@@ -57,9 +57,9 @@ describe EstimationDecorator do
 
   describe '#estimation_items' do
     it 'orders items by created_at' do
-      estimation = FactoryGirl.create(:estimation)
-      new_item = FactoryGirl.create :estimation_item, estimation: estimation, created_at: 1.minute.ago
-      old_item = FactoryGirl.create :estimation_item, estimation: estimation, created_at: 1.hour.ago
+      estimation = FactoryBot.create(:estimation)
+      new_item = FactoryBot.create :estimation_item, estimation: estimation, created_at: 1.minute.ago
+      old_item = FactoryBot.create :estimation_item, estimation: estimation, created_at: 1.hour.ago
 
       expect(estimation.decorate.estimation_items).to eq [old_item, new_item]
     end
@@ -67,8 +67,8 @@ describe EstimationDecorator do
 
   describe '#buffer_health' do
     it 'outputs percentage' do
-      estimation = FactoryGirl.create(:estimation)
-      new_item = FactoryGirl.create :estimation_item, estimation: estimation, value: 1, actual_value: 2
+      estimation = FactoryBot.create(:estimation)
+      new_item = FactoryBot.create :estimation_item, estimation: estimation, value: 1, actual_value: 2
 
       expect(estimation.decorate.buffer_health).to eq '100%'
     end
@@ -76,7 +76,7 @@ describe EstimationDecorator do
 
   # TODO Make smarter function here. Buffer health 1.1 is norm at the beginning of the project and critically wrong at the and
   describe '#buffer_health_class' do
-    subject { FactoryGirl.create(:estimation) }
+    subject { FactoryBot.create(:estimation) }
 
     it 'is :bg-success when health is less than 0.8' do
       expect(subject).to receive(:buffer_health).and_return(0.0)
