@@ -2,7 +2,7 @@ require 'rails_helper'
 
 RSpec.describe "estimations/index.html.erb", :type => :view do
   it 'should contain user\'s private estimations' do
-    estimations = (1..3).map { FactoryGirl.create(:estimation) }
+    estimations = (1..3).map { FactoryBot.create(:estimation) }
     assign(:estimations, estimations)
 
     render
