@@ -1,57 +1,55 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your
-# database schema. If you need to create the application database on another
-# system, you should be using db:schema:load, not running all the migrations
-# from scratch. The latter is a flawed and unsustainable approach (the more migrations
-# you'll amass, the slower it'll run and the greater likelihood for issues).
+# This file is the source Rails uses to define your schema when running `rails
+# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20150821230438) do
+ActiveRecord::Schema.define(version: 2015_07_01_170544) do
 
   create_table "estimation_items", force: :cascade do |t|
-    t.integer  "value",         default: 0
-    t.string   "title"
-    t.integer  "estimation_id"
-    t.datetime "created_at",                    null: false
-    t.datetime "updated_at",                    null: false
-    t.boolean  "fixed",         default: false, null: false
-    t.integer  "quantity",      default: 1,     null: false
-    t.float    "actual_value"
+    t.integer "value"
+    t.string "title"
+    t.integer "estimation_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.boolean "fixed", default: false, null: false
+    t.integer "quantity", default: 1, null: false
+    t.float "actual_value"
+    t.index ["estimation_id"], name: "index_estimation_items_on_estimation_id"
   end
-
-  add_index "estimation_items", ["estimation_id"], name: "index_estimation_items_on_estimation_id"
 
   create_table "estimations", force: :cascade do |t|
-    t.string   "title"
-    t.integer  "user_id"
-    t.datetime "created_at",                    null: false
-    t.datetime "updated_at",                    null: false
-    t.boolean  "tracking_mode", default: false, null: false
+    t.string "title"
+    t.integer "user_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.boolean "tracking_mode", default: false, null: false
+    t.index ["user_id"], name: "index_estimations_on_user_id"
   end
-
-  add_index "estimations", ["user_id"], name: "index_estimations_on_user_id"
 
   create_table "users", force: :cascade do |t|
-    t.string   "email",               default: "", null: false
+    t.string "email", default: "", null: false
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",       default: 0,  null: false
+    t.integer "sign_in_count", default: 0, null: false
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
-    t.string   "current_sign_in_ip"
-    t.string   "last_sign_in_ip"
+    t.string "current_sign_in_ip"
+    t.string "last_sign_in_ip"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "provider"
-    t.string   "uid"
+    t.string "provider"
+    t.string "uid"
+    t.index ["email"], name: "index_users_on_email", unique: true
+    t.index ["provider"], name: "index_users_on_provider"
+    t.index ["uid"], name: "index_users_on_uid"
   end
 
-  add_index "users", ["email"], name: "index_users_on_email", unique: true
-  add_index "users", ["provider"], name: "index_users_on_provider"
-  add_index "users", ["uid"], name: "index_users_on_uid"
-
+  add_foreign_key "estimation_items", "estimations"
+  add_foreign_key "estimations", "users"
 end
